+++ tweaks/tweaks.js
@@ -34,7 +34,7 @@
 
     /* --- Version marker (stamped by the publish script) ------------------ */
 
-    var TWEAKS_VERSION = '1.24 b21 04 Aug 19:51';
+    var TWEAKS_VERSION = '1.25 b22 04 Aug 19:56';
     var SHOW_BADGE = true;
 
     function showVersionBadge() {